One-for-All-125 - Added the DeleteFieldButton.
