+++ apps/web/src/feature/app/field/AddAppFieldContainer.tsx
@@ -55,6 +55,7 @@ export const AddAppFieldContainer = ({ index }: { index: number }) => {
       isOpen={isOpen}
       setIsOpen={setIsOpen}
       saveHandler={saveHandler}
+      deleteHandler={() => undefined}
     />
   );
 };
