+++ apps/web/src/feature/app/field/ModifyAppFieldContainer.tsx
@@ -80,6 +80,7 @@ export const ModifyAppFieldContainer = ({
       isOpen={isOpen}
       setIsOpen={setIsOpen}
       saveHandler={saveHandler}
+      deleteHandler={() => undefined}
     />
   );
 };
